Свел поиск комнат по критерию к одному HQL запросу с подзапросом

+++ hotel/trunk/src/main/java/ru/yandex/zhmyd/hotel/repository/dao/OrderDao.java
@@ -4,7 +4,7 @@ import ru.yandex.zhmyd.hotel.repository.entity.RoomEntity;
 import ru.yandex.zhmyd.hotel.repository.entity.OrderEntity;
 import ru.yandex.zhmyd.hotel.repository.entity.UserEntity;
 
-public interface OrderDao extends GenericDao<OrderEntity, Integer> {
-    public UserEntity getUserByOrderId(Integer roomOrderId);
-    public RoomEntity getRoomByOrderId(Integer roomOrderId);
+public interface OrderDao extends GenericDao<OrderEntity, Long> {
+    public UserEntity getUserByOrderId(Long roomOrderId);
+    public RoomEntity getRoomByOrderId(Long roomOrderId);
 }
